# test_models: speed up test