Speeds up 214 runs through this function from 73 seconds to 61 seconds. With a baseline of ~30 seconds to do the setUp and tearDown, this now takes about 70% of the previous time taken for just this function

- 214 function runs (`FILEREADER_CACHE=1`, `-j50`)
  - Before: 73s
  - After: 61ss
  - Baseline (only setUpClass, setUp, tearDown): 30s
  - Speedup: ~27%
- 691 function runs (`FILEREADER_CACHE=1`, `-j50`)
  - Before: 190s
  - After: 153s
  - Baseline (only setUpClass, setUp, tearDown): 70s
  - Speedup: ~30%